+++ packages/arcade_cache/lib/arcade_cache.dart
@@ -1,2 +1,4 @@
 export 'src/cache_manager.dart' show BaseCacheManager;
 export 'src/memory_cache_manager.dart' show MemoryCacheManager;
+export 'src/pubsub_event.dart'
+    show PubSubEvent, PubSubMessage, PubSubSubscribed, PubSubUnsubscribed;